pyproject.toml
[tool.poetry]
name = "trainee-tt"
version = "0.1.0"
description = ""
authors = ["Evgeny Korolev <evgeny.korolev555@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
